imos_backup: use su for restore operation
if the homedir is not set, it will be /root. then when the backups user
will try to execute commands it might fail. an example for a failed
command is trying to run `find -exec ...`, that will fail with a
very vague error of `find: pred.c:1932: launch: Assertion 'starting_desc
>= 0' failed.`. using su instead of sudo will also move users to their
home directories, preventing the restore from failing