imos_apache2: ProxyPass retry=0
prevent webapps from timing out for up to 1 minute (default) when they are
being restarted. this is due to apache stopping to forward requests for that 1
minute period if the last time it accessed the webapp it was down. retry=0
means apache will not try to remember any state about the webapp and will just
forward any request to it