imos_webapps/imos_artifacts: parallel deploy support
make use of tomcat 7 parallel deployment if configured to. enabled
by specifying `"parallel_deploy": true` on the tomcat container.
this is container wide as it needs to enable it on the whole tomcat
installation.

this change will not affect current operation of running tomcat7
containers!

the working behind this is simple - a version (calculated from the
mtime of the downloaded cached file) is calculated and appended to
the deployed war file, s.t. app.war will deploy to app##version.war
and app directory will become app##version directory. this results
in tomcat deploying the newly versioned webapp without needing a
restart