Change to Vagrantfile to allow multi-machine commands to work
e.g. previously, `vagrant status`, which is useful for quickly seeing the status of all machines, would fail because of 8afa926b3fa1d.
This acheives (I believe) the same as intended in 8afa926b3fa1d, but allows multi-machine commands to still work as expected.

As a bonus, using out-of-the box vagrant config instead of custom code.

As a note, a `vagrant provision` (with no node name supplied) will be applied to any 'up' nodes - but I think this is probably ok.