imos_postgresql: cache after resources creation
imos_postgresql creates roles and database caches in the chef
cache directory, that is to speed up chef provisioning because
usually nothing is updated. this commit makes sure the cache changes
persist to disk only after the roles and databases resources were
created. pushing it to a ruby_block ensures it is called as a
resource after the other resources were successfully evaluated