talend: talend-trigger can run in bulk
factor talend-trigger to be able to run on a bunch of files at the
same time, that is by traversing the list of harvesters and
matching files, rather than traversing the list of files and
matching harvesters. also significantly improve command line
usability of talend-trigger. support 'noop' (no operation) mode
so it is easier to debug what is going to happen