'Pin' plugins
This has the effect of allowing plugins which are otherwise bundled in the jenkins.war to be upgraded (otherwise, they are upgraded but then nuked with the bundled version when jenkins is restarted).

The upgrade (of certain plugins) is necessary otherwise cyclic dependencies exists which in turn causes problems downstream (i.e. some jobs don't have the plugins they need to run).