talend: upper limit for talend-trigger operation
Need limits in life. Prevent situations when talend-trigger is
triggered with many files (100K and more), trying to allocate big
arrays and also use a lot of temporary space in temp directories.
This commit sets a default 2048 items chunk, which will make the
talend-trigger process up to that amount in every talend
operation.