imos_po: celery has queue per task
Factor celery to use a queue per task. That way celery can be
configured to have a limit of threads per task, which will be 1 by
default.

Must be don in a ruby_block so it happens after the data-services
git repository is updated and it actually creates all the defined
tasks.