imos_po: Vagrant trigger celeryd job creation
When on vagrant (dev), always trigger celeryd job creation. This
is due to the `tasks.py` file living outside of the chef box. In
the scenario of someone destroying a box and then running
`vagrant up` they might end up with nothing in `/etc/supervisor.d`
if `tasks.py` hasn't been modified.