imos_po: Create watches on first provision
The problem this PR solves is described in #240.

The fix is rather awkward, but works properly. Any related
template would subscribe to be notified after git[data_services]
is updated, however it would evaluate the list of watches before
the git repository was updated as a template variables.
Eventually when the template is evaluated, it would use a stale
watch list which consists of watches *before* git[data_services]
was updated. The template however would be evaluated after
git[data_services] was updated.

The fix moved the evaluation of the watch list to the template
itself. This causes the list of watches to be evaluated *after*
git[data_services] was updated, thus resulting in the right list
of watches used by the template.